Route, View
Route: Actually, it is the path of the project. After typing the route name with the URL, we can go to our expected page. There are many types of routes type: Get, Post, Delete, Put, etc. We can use them when we need those types of work.

View: The view file format will be blade.php. Here, we can use HTML and other things. We use the view to show route content. We can return a string and inline HTML as the return of the route.